requirements.txt
fastapi
uvicorn
python-multipart
moviepy
librosa
numpy
openai
google-generativeai